+++ AirCasting/AirCastingApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct AirCastingApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            Dashboard()
         }
     }
 }
