+++ AirCasting/Extensions/Color.swift
@@ -1,0 +1,18 @@
+//
+//  Color.swift
+//  AirCasting
+//
+//  Created by Lunar on 07/01/2021.
+//
+
+import Foundation
+import SwiftUI
+
+extension Color {
+    static var aircastingGray: Color {
+        return Color("AircastingGray700")
+    }
+    static var darkBlue: Color {
+        return Color("DarkBlue")
+    }
+}
